feat: update exhibit-step screens navigation
